Cogstate-RFI-RFP/server/requirements.txt
markitdown[all]
unstructured[all]
unstructured[doc]
unstructured
unstructured-inference
pdfminer.six
pypdf
pymupdf
onnxruntime
python-dotenv
tqdm
SQLAlchemy
langchain_openai
langchain
pydantic
google-generativeai
langchain_google_genai
chromadb
langchain_chroma
fastapi
uvicorn[standard]
python-multipart
# Additional dependencies for enhanced functionality
python-docx
openpyxl
pandas
langchain-text-splitters
google-cloud-storage